+++ src/system/fs/stat.ts
